requirements.txt
streamlit>=1.30.0
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
openai>=1.12.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
jinja2>=3.1.2
